+++ example/lib/helper/colours.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 class Colours {
   Colours._();
 
-  static Color blue() => const Color(0xff5e6ceb);
+  static Color blue = const Color(0xff5e6ceb);
 
-  static Color blueDark() => const Color(0xff4D5DFB);
+  static Color blueDark = const Color(0xff4D5DFB);
 }